+++ packages/serverless/index.ts
@@ -5,4 +5,4 @@
  * The software is provided "as is", without warranty of any kind.
  * ######## */
 
-export * from './src/aws_Lambda';
+export * from './src/awsLambda';
