+++ packages/serverless/src/awsLambda.ts
